mnist model

+++ xdsl/transforms/convert_onnx_to_linalg.py
@@ -62,12 +62,12 @@ class AddOpLowering(RewritePattern):
 class ReluOpLowering(RewritePattern):
     @op_type_rewrite_pattern
     def match_and_rewrite(self, relu: onnx.Relu, rewriter: PatternRewriter, /):
-        body = Region(Block(arg_types=(f64, f64)))
+        body = Region(Block(arg_types=(f32, f32)))
         affine_map = AffineMapAttr(AffineMap.from_callable(lambda d0, d1: (d0, d1)))
         rewriter.replace_matched_op(
             (
                 empty := tensor.EmptyOp((), relu.res.type),
-                zero := arith.Constant(FloatAttr(0, f64)),
+                zero := arith.Constant(FloatAttr(0, f32)),
                 linalg.Generic(
                     (relu.operand,),
                     (empty.tensor,),
